% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2group2function.R
\name{boxplot_function}
\alias{boxplot_function}
\title{Generate Boxplot by DRG Code}
\usage{
boxplot_function(data, payment_type)
}
\arguments{
\item{data}{Data frame containing DRG data.}

\item{payment_type}{A string indicating the type of payment to plot: "medicare", "total", or "covered".}
}
\value{
A ggplot object of the boxplot.
}
\description{
This function generates a boxplot of payments by DRG code.
Users can choose between average Medicare payments, average total payments, or average covered charges.
}
\examples{
boxplot_function(drg_data, "medicare")
}
